requirements.txt
asgiref==3.5.0
Django==4.0.2
python-crontab==2.6.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
